Added batch size cli arg

+++ src/file_io/results_ser.rs
@@ -28,7 +28,7 @@ pub struct TrainingResultsSer {
     elapsed_time: f32,
     /// Total number of iterations until the
     /// network was considered fully trained
-    total_epochs: u64,
+    total_epochs: usize,
     /// Predicted values from feeding validtion
     /// set inputs into the trained network
     predicted_output: Array2<f64>,
@@ -41,7 +41,7 @@ impl TrainingResultsSer {
         metric_value: f64,
         metric_passed: bool,
         elapsed_time: f32,
-        total_epochs: u64,
+        total_epochs: usize,
         predicted_output: Array2<f64>,
     ) -> Self {
         let metric: MetricSer = MetricSer {
